refactor(blockchain): Implement cloneable readers (#359)
