+++ blockchain/src/in_memory.rs
@@ -1,13 +1,34 @@
-use crate::{block::BlockMemory, transaction::TransactionMemory};
+use crate::{
+    block::BlockMemory,
+    transaction::{TransactionMemory, TransactionMemoryReader},
+};
 
-#[derive(Debug, Default)]
+#[derive(Debug)]
+pub struct SharedMemoryReader {
+    pub block_memory: BlockMemory,
+    pub transaction_memory: TransactionMemoryReader,
+}
+
+impl SharedMemoryReader {
+    pub fn new(block_memory: BlockMemory, transaction_memory: TransactionMemoryReader) -> Self {
+        Self {
+            block_memory,
+            transaction_memory,
+        }
+    }
+}
+
+#[derive(Debug)]
 pub struct SharedMemory {
     pub block_memory: BlockMemory,
     pub transaction_memory: TransactionMemory,
 }
 
 impl SharedMemory {
-    pub fn new() -> Self {
-        Self::default()
+    pub fn new(block_memory: BlockMemory, transaction_memory: TransactionMemory) -> Self {
+        Self {
+            block_memory,
+            transaction_memory,
+        }
     }
 }
