+++ blockchain/src/transaction/in_memory.rs
@@ -1,31 +1,19 @@
-use {
-    crate::{
-        in_memory::SharedMemory,
-        transaction::{
-            ExtendedTransaction, TransactionQueries, TransactionRepository, TransactionResponse,
-        },
-    },
-    moved_shared::primitives::B256,
-    std::{collections::HashMap, convert::Infallible},
-};
+use {crate::transaction::ExtendedTransaction, moved_shared::primitives::B256};
 
-#[derive(Debug, Default)]
-pub struct TransactionMemory {
-    transactions: HashMap<B256, ExtendedTransaction>,
+#[derive(Debug, Clone)]
+pub struct TransactionMemoryReader {
+    transactions: evmap::ReadHandle<B256, Box<ExtendedTransaction>>,
 }
 
-impl TransactionMemory {
-    pub fn new() -> Self {
-        Self::default()
-    }
-
-    pub fn extend(&mut self, tx: impl IntoIterator<Item = ExtendedTransaction>) {
-        self.transactions
-            .extend(tx.into_iter().map(|tx| (tx.hash(), tx)));
+impl TransactionMemoryReader {
+    pub fn new(transactions: evmap::ReadHandle<B256, Box<ExtendedTransaction>>) -> Self {
+        Self { transactions }
     }
 
     pub fn by_hash(&self, hash: B256) -> Option<ExtendedTransaction> {
-        self.transactions.get(&hash).cloned()
+        self.transactions
+            .get(&hash)
+            .and_then(|v| v.iter().next().map(|v| *v.clone()))
     }
 
     pub fn by_hashes(&self, hashes: impl IntoIterator<Item = B256>) -> Vec<ExtendedTransaction> {
@@ -36,81 +24,31 @@ impl TransactionMemory {
     }
 }
 
-#[derive(Debug, Default)]
-pub struct InMemoryTransactionRepository;
-
-impl InMemoryTransactionRepository {
-    pub fn new() -> Self {
-        Self
-    }
+#[derive(Debug)]
+pub struct TransactionMemory {
+    transactions: evmap::WriteHandle<B256, Box<ExtendedTransaction>>,
 }
 
-#[derive(Debug, Default)]
-pub struct InMemoryTransactionQueries;
-
-impl InMemoryTransactionQueries {
-    pub fn new() -> Self {
-        Self
-    }
-}
-
-impl TransactionRepository for InMemoryTransactionRepository {
-    type Err = Infallible;
-    type Storage = SharedMemory;
-
-    fn extend(
-        &mut self,
-        storage: &mut Self::Storage,
-        transactions: impl IntoIterator<Item = ExtendedTransaction>,
-    ) -> Result<(), Self::Err> {
-        storage.transaction_memory.extend(transactions);
-        Ok(())
-    }
-}
-
-impl TransactionQueries for InMemoryTransactionQueries {
-    type Err = Infallible;
-    type Storage = SharedMemory;
-
-    fn by_hash(
-        &self,
-        storage: &Self::Storage,
-        hash: B256,
-    ) -> Result<Option<TransactionResponse>, Self::Err> {
-        Ok(storage
-            .transaction_memory
-            .by_hash(hash)
-            .map(TransactionResponse::from))
-    }
-}
-
-#[cfg(any(feature = "test-doubles", test))]
-mod test_doubles {
-    use super::*;
-
-    impl TransactionQueries for () {
-        type Err = Infallible;
-        type Storage = ();
-
-        fn by_hash(
-            &self,
-            _: &Self::Storage,
-            _: B256,
-        ) -> Result<Option<TransactionResponse>, Self::Err> {
-            Ok(None)
-        }
+impl TransactionMemory {
+    pub fn new(transactions: evmap::WriteHandle<B256, Box<ExtendedTransaction>>) -> Self {
+        Self { transactions }
     }
 
-    impl TransactionRepository for () {
-        type Err = Infallible;
-        type Storage = ();
+    pub fn extend(&mut self, tx: impl IntoIterator<Item = ExtendedTransaction>) {
+        self.transactions
+            .extend(tx.into_iter().map(|tx| (tx.hash(), Box::new(tx))));
+    }
 
-        fn extend(
-            &mut self,
-            _: &mut Self::Storage,
-            _: impl IntoIterator<Item = ExtendedTransaction>,
-        ) -> Result<(), Self::Err> {
-            Ok(())
-        }
+    pub fn by_hash(&self, hash: B256) -> Option<ExtendedTransaction> {
+        self.transactions
+            .get(&hash)
+            .and_then(|v| v.iter().next().map(|v| *v.clone()))
+    }
+
+    pub fn by_hashes(&self, hashes: impl IntoIterator<Item = B256>) -> Vec<ExtendedTransaction> {
+        hashes
+            .into_iter()
+            .filter_map(|hash| self.by_hash(hash))
+            .collect()
     }
 }
