+++ blockchain/src/transaction/mod.rs
@@ -1,7 +1,7 @@
 pub use {
-    in_memory::{InMemoryTransactionQueries, InMemoryTransactionRepository, TransactionMemory},
-    read::{TransactionQueries, TransactionResponse},
-    write::{ExtendedTransaction, TransactionRepository},
+    in_memory::{TransactionMemory, TransactionMemoryReader},
+    read::{TransactionQueries, TransactionResponse, in_memory::InMemoryTransactionQueries},
+    write::{ExtendedTransaction, TransactionRepository, in_memory::InMemoryTransactionRepository},
 };
 
 mod in_memory;
